requirements.txt
rich>=13.3.0
questionary>=1.10.0
jinja2>=3.1.2
pyyaml>=6.0
